REFACTOR move Filters to own template
