+++ pwa/src/templates/components/ComponentsTemplate.tsx
@@ -1,28 +1,20 @@
 import * as React from "react";
 import * as styles from "./ComponentsTemplate.module.css";
 import * as _ from "lodash";
-import { Button, FormField, FormFieldInput, FormFieldLabel, Heading2 } from "@gemeente-denhaag/components-react";
+import { Button, Heading2 } from "@gemeente-denhaag/components-react";
 import { t } from "i18next";
-import { useForm } from "react-hook-form";
-import { InputText, Container, SelectMultiple } from "@conduction/components";
+import { Container } from "@conduction/components";
 import { ComponentResultTemplate, TComponentResultsLayout } from "../componentResult/ComponentResultsTemplate";
 import { FiltersContext } from "../../context/filters";
 import { getFilteredComponents } from "../../services/getFilteredComponents";
 import { useComponent } from "../../hooks/components";
 import { QueryClient } from "react-query";
+import { FiltersTemplate } from "../templateParts/filters/FiltersTemplate";
 
 export const ComponentsTemplate: React.FC = () => {
-  const [filters, setFilters] = React.useContext(FiltersContext);
+  const [filters] = React.useContext(FiltersContext);
   const [filteredComponents, setFilteredComponents] = React.useState<any[]>([]);
   const [display, setDisplay] = React.useState<TComponentResultsLayout>("table");
-
-  const {
-    register,
-    watch,
-    reset,
-    control,
-    formState: { errors },
-  } = useForm();
 
   const queryClient = new QueryClient();
   const _useComponent = useComponent(queryClient);
@@ -31,18 +23,8 @@ export const ComponentsTemplate: React.FC = () => {
   React.useEffect(() => {
     if (!getComponents.isSuccess) return;
 
-    reset({ name: filters.name, layers: filters.layers?.map((t) => getSelectObjectFromValue(t)) });
-
     setFilteredComponents(getFilteredComponents(getComponents.data, filters));
   }, [filters, getComponents.isSuccess]);
-
-  React.useEffect(() => {
-    const subscription = watch(({ name, layers }) => {
-      setFilters({ name: name, layers: layers?.map((t: any) => t.value) });
-    });
-
-    return () => subscription.unsubscribe();
-  });
 
   return (
     <Container layoutClassName={styles.container}>
@@ -67,26 +49,7 @@ export const ComponentsTemplate: React.FC = () => {
         </div>
       </div>
 
-      <form className={styles.form}>
-        <FormField>
-          <FormFieldInput>
-            <FormFieldLabel>Filter op naam</FormFieldLabel>
-            <InputText name="name" validation={{ required: true }} {...{ errors, register }} />
-          </FormFieldInput>
-        </FormField>
-
-        <FormField>
-          <FormFieldInput>
-            <FormFieldLabel>Filter op laag</FormFieldLabel>
-            <SelectMultiple
-              defaultValue={filters.layers?.map((f) => getSelectObjectFromValue(f))}
-              name="layers"
-              options={layers}
-              {...{ errors, control, register }}
-            />
-          </FormFieldInput>
-        </FormField>
-      </form>
+      <FiltersTemplate />
 
       {filteredComponents.length > 0 && <ComponentResultTemplate results={filteredComponents} type={display} />}
 
@@ -94,15 +57,3 @@ export const ComponentsTemplate: React.FC = () => {
     </Container>
   );
 };
-
-const getSelectObjectFromValue = (value: string | undefined): any | undefined => {
-  return layers.find((t) => t.value === value);
-};
-
-const layers = [
-  { label: "Interactie", value: "interactie" },
-  { label: "Proces", value: "proces" },
-  { label: "Integratie", value: "integratie" },
-  { label: "Services", value: "services" },
-  { label: "Data", value: "data" },
-];
