+++ pwa/src/templates/templateParts/filters/FiltersTemplate.module.css
@@ -1,0 +1,5 @@
+.form {
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  gap: var(--nlportal-space-inline-lg);
+}
